requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
motor==3.6.0
python-dotenv==0.19.0